Restructuring and implementing idx to id and id to idx
